Use background image instead of <img> as placeholder. So it doesn't overlap with the height calculation and the fixed footer.
