+++ src/App.css
@@ -26,6 +26,13 @@
   padding: '0.5em 0';
 }
 
+.disconnected {
+  background-image: url('./assets/disconnected.png');
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: 600px 450px;
+}
+
 .App-footer {
   align-items: center;
   justify-content: center;
